assetList as reusable component
