+++ src/presentation/components/button-list.tsx
@@ -3,19 +3,17 @@ import React from 'react';
 interface Props {
   children?: React.ReactElement | React.ReactElement[];
   title: string;
-  type: 'assets' | 'transactions';
+  emptyText: string;
 }
 
-const ButtonList: React.FC<Props> = ({ children, title, type }: Props) => {
+const ButtonList: React.FC<Props> = ({ children, title, emptyText }: Props) => {
   return (
     <div>
       <h2 className="mt-2 text-lg font-medium text-left text-white">{title}</h2>
       <div className="h-64 overflow-y-auto" id="btn-list">
         <div className="py-4 space-y-4">
           {React.Children.count(children) === 0 ? (
-            <span className="text-sm font-medium text-white">
-              Your transactions will appear here
-            </span>
+            <span className="text-sm font-medium text-white">{emptyText}</span>
           ) : (
             children
           )}
